Tidy up and add readme
