+++ projects/plugins/jetpack/package.json
@@ -35,6 +35,7 @@
 		"clean-extensions": "rm -rf _inc/blocks/ ",
 		"clean-search": "rm -rf _inc/build/instant-search/* ",
 		"distclean": "rm -rf node_modules && yarn clean",
+		"fixtures:regenerate": "cross-env REGENERATE_FIXTURES=y GENERATE_MISSING_FIXTURES=y yarn jest",
 		"fixtures:generate": "cross-env GENERATE_MISSING_FIXTURES=y yarn jest",
 		"install-if-deps-outdated": "yarn install --check-files --production=false --frozen-lockfile",
 		"test-adminpage": "yarn test-client && yarn test-gui",
